requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==1.10.13
openai==1.2.4
python-dotenv==1.0.0
PyPDF2==3.0.1
python-multipart==0.0.6
email-validator==2.0.0
pymongo==4.6.1
bson==0.5.10
motor==3.3.1
numpy==1.24.3 